Re-support canonical link
We do this via JavaScript to avoid issues with plugin load order; the
canonical link might not have been set when it's our turn.  Of course,
that doesn't work if JavaScript is not supported, but that's likely
not much of an issue, because (most) social networks require JS anyway,
and they likely are going to look back on the page to find out more,
and can grab the canonical link from there.